requirements.txt
pip>=24.0.0
pylint==3.0.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
#ipython==8.13.0

streamlit==1.37.0
pandas==2.2.2
watchdog==4.0.0
plotly==5.18.0
openpyxl==3.1.2

transformers==4.44.0
datasets
numpy==1.26.4
torch
scikit-learn

psycopg2-binary==2.9.9

pyperclip